refactor(exercise): move exercise type out of widgets
