+++ lib/presentation/widgets/exercise_selector.dart
@@ -1,24 +1,6 @@
-import 'package:fitness_counter/fitness_counter.dart';
 import 'package:flutter/material.dart';
 
-/// Exercise types available for rep counting.
-enum ExerciseType {
-  lateralRaise('Lateral Raise'),
-  singleSquat('Single Squat');
-
-  final String displayName;
-  const ExerciseType(this.displayName);
-
-  /// Get the configuration for this exercise type.
-  ExerciseConfig get config {
-    switch (this) {
-      case ExerciseType.lateralRaise:
-        return ExerciseConfigs.lateralRaise;
-      case ExerciseType.singleSquat:
-        return ExerciseConfigs.singleSquat;
-    }
-  }
-}
+import '../../domain/models/exercise_type.dart';
 
 /// Dropdown widget for selecting an exercise.
 class ExerciseSelectorDropdown extends StatelessWidget {
